Expose the symbol property "introduced"
When outputting a CSV table with symbols in mode "clausify",
add the column "introduced" and fill it with the property
`introduced()`.
This property signifies whether the symbol was introduced during
clausification. It has the value 1 for Tseitin and Skolem symbols
and the value 0 for other symbols. In function symbols its value
coincides with the property "skolem".